@@ -1,36 +1,58 @@
-const express = require('express');
-const sqlite3 = require('sqlite3').verbose();
+const express = require("express")
+const cors = require("cors")
+const bodyParser = require("body-parser")
+const cookieParser = require("cookie-parser")
+const passport = require("passport")
 
-const app = express();
-app.use(express.json());
-let db = new sqlite3.Database('./db/main.db');
+if (process.env.NODE_ENV !== "production") {
+    // load env variables from .env file in non prod
+    // TODO: figure out how to do this in prod
+    require('dotenv').config()
+}
 
-app.get('/api/data', (req, res) => {
-    db.all('SELECT * FROM main', (err, rows) => {
-        res.json(rows);
-    });
-});
+require("./utils/connectdb")
 
-app.post('/api/data', (req, res) => {
-    console.log(req.body);
-    let MILES = req.body.MILES;
-    let GAS = req.body.GAS;
-    let PRICE_PER_GAL = req.body.PRICE_PER_GAL;
-    if (typeof MILES !== "number" || typeof GAS !== "number" || typeof PRICE_PER_GAL !== "number") {
-        res.sendStatus(400);
-    }
-    else {
-        db.run("INSERT INTO main VALUES(?, ?, ?)", MILES, GAS, PRICE_PER_GAL, (err) => {
-            if (err) {
-                console.log(err);
-                res.sendStatus(500);
-            }
-            else {
-                res.sendStatus(200);
-            }
-        });
-    }
-});
+require("./strategies/JwtStrategy")
+require("./strategies/LocalStrategy")
+require("./authenticate")
 
+const userRouter = require("./routes/userRoutes")
+const dataRouter = require('./routes/dataRoutes')
 
-app.listen(5000, () => console.log('Example app running on port 5000'));
+const app = express()
+
+app.use(bodyParser.json())
+app.use(cookieParser(process.env.COOKIE_SECRET))
+
+//add client url to cors policy
+const whitelist = process.env.WHITELISTED_DOMAINS
+    ? process.env.WHITELISTED_DOMAINS.split(",")
+    : []
+
+const corsOptions = {
+    origin: function(origin, callback) {
+        if (!origin || whitelist.indexOf(origin) !== -1) {
+            callback(null, true)
+        } else {
+            callback(new Error("Not allowed by CORS"))
+        }
+    },
+    credentials: true,
+}
+
+app.use(cors(corsOptions))
+
+app.use(passport.initialize())
+
+app.use("/users", userRouter)
+app.use("/data", dataRouter)
+
+app.get("/", function(req, res) {
+    res.send({status: "success"})
+})
+
+//start server on env.PORT or 8081
+const server = app.listen(process.env.PORT || 8081, function() {
+    const port = server.address().port
+    console.log(`App started on port ${port}`)
+})
